Windows bugfixes
On Windows getting a list of displays resulted in a terminal window
being briefly displayed. If the reason for getting a list of displays
was a screen resolution change this resulted in another screen change
due to the terminal window popping up. Which resulted in an endless
loop.

This fix hides the terminal window.